app/backend/requirements.txt
# Core dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.5.0,<2.11
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# ASGI
python-multipart>=0.0.6

# Development and testing
pytest>=8.4.1
pytest-asyncio>=1.1.0
httpx>=0.27.0

# Crypto
cryptography>=42.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0

# AI Drivers
openai>=1.0.0
sse-starlette>=1.6.0

# Utilities
tenacity>=8.0.0
